disabling sendfile in nginx
Sendfile is broken in virtual box shared folders, and produces garbage data.  This triggers a bug where nginx will only host the current static file the first time it is requested.  if the static file is change, for example if you are working on javascript files, it won't show the updates, and garbage data get's in the stream, usually at the end.